+++ src/burn/drv/pre90s/d_dday.cpp
@@ -39,7 +39,7 @@ static UINT8 DrvDips[2];
 static UINT8 DrvInputs[1];
 static UINT8 DrvReset;
 
-static INT32 DrvAnalogPort0 = 0;
+static INT16 DrvAnalogPort0 = 0;
 
 #define A(a, b, c, d) {a, b, (UINT8*)(c), d}
 static struct BurnInputInfo DdayInputList[] = {
